feat(conf/gifs/participants): created ParticipantsPaneFooter and updated comments
